Added custom action upload_user_image to the UserViewSet
Introduced a custom action upload_user_image within the UserViewSet to enable users to upload images for specific user instances.
Implemented the get_serializer_class method to dynamically select the serializer class based on the action being performed. Returns UserImageSerializer for the upload_user_image action and UserSerializer for other actions.
Decorated the upload_user_image method with @action to define a custom endpoint for uploading user images. It handles POST requests and operates on a single object instance.
Upon receiving a valid request, the method saves the uploaded image data associated with the user instance and returns a response with status code 200 along with the serialized data.
If the request data is invalid, the method returns a response with status code 400 and includes the serializer errors.